prints the first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 int main(void)
 {
 	int i;
-	unsigned long long int sum1 = 0, sum2 = 1;
-	unsigned long long int fib;
+	unsigned long int sum1 = 0, sum2 = 1;
+	unsigned long int fib;
 
 	for (i = 0; i  < 98; i++)
 	{
 		fib = sum1 + sum2;
 		sum1 = sum2;
 		sum2 = fib;
-		printf("%llu", fib);
+		printf("%lu", fib);
 		if (i == 97)
 			break;
 		printf(", ");
